% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular.R
\name{circular}
\alias{circular}
\title{Generate circular weigthing matrices}
\usage{
circular(nrow, ncol, ab)
}
\arguments{
\item{nrow}{Number of rows}

\item{ncol}{Number of columns}

\item{ab}{Ahead - behind}
}
\value{
An object of class \code{nb}
}
\description{
The function generates k-ahead and k-behind weighting matrices as in Kelejian and Prucha (1999)
}
\details{
Generate circular weigthing matrices
}
\examples{
ab.3 <- circular(10, 10, 3)
ab.5 <-  circular(15, 15, 5)
}
\author{
Gianfranco Piras \email{gpiras@mac.com}
}
\keyword{spatial}
