% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coldis.R
\docType{data}
\name{coldis}
\alias{coldis}
\alias{knn10columbus}
\title{Object of class distance for Columbus dataset 
10-nearest neighbors matrix for columbus dataset}
\format{
A list of neighbors and distances for the columbus neighborhoods
}
\description{
Object of class distance for Columbus dataset 
10-nearest neighbors matrix for columbus dataset
}
\author{
Gianfranco Piras \email{gpiras@mac.com}
}
