% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.impacts.sphet.R
\name{impacts}
\alias{impacts}
\title{Generate impacts for spreg lag and sarar models}
\usage{
impacts(obj, mvn = TRUE, n_mvn = 3000, weg = NULL)
}
\arguments{
\item{obj}{An object of class sphet}

\item{mvn}{Inference based on MC simulation of mvn samples}

\item{n_mvn}{Number of samples}

\item{weg}{Eigenvalues can be provided by the user}
}
\value{
Estimate of the Average Total, Average Direct, and Average Indirect Effects
}
\description{
Generate impacts for spreg lag and sarar models
}
\examples{
data(columbus, package="spdep")
listw <- spdep::nb2listw(col.gal.nb)
res <- spreg(CRIME~HOVAL + INC, data=columbus , listw= listw,
            het = TRUE, verbose = FALSE, model = "sarar")
summary(res)
effects <- impacts(res, n_mvn = 3000)
summary(effects)
}
