% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.impacts.sphet.R
\name{impacts.sarar}
\alias{impacts.sarar}
\title{Generate impacts}
\usage{
\method{impacts}{sarar}(
  object,
  n_mvn = 3000,
  weg = NULL,
  inference = FALSE,
  exact = TRUE,
  m = 30,
  trW = NULL,
  mvn = TRUE
)
}
\arguments{
\item{object}{An object of class sphet}

\item{n_mvn}{Number of samples}

\item{weg}{Eigenvalues can be provided by the user}

\item{inference}{should inference be performed}

\item{exact}{should inference be exact}

\item{m}{number of traces of w - default is 30}

\item{trW}{either NULL or a trW from spatialreg}

\item{mvn}{Inference based on MC simulation of mvn samples}
}
\value{
Estimate of the Average Total, Average Direct, and Average Indirect Effects
}
\description{
Generate impacts
}
\examples{
data(columbus, package="spdep")
listw <- spdep::nb2listw(col.gal.nb)
res <- spreg(CRIME~HOVAL + INC, data=columbus , listw= listw,
            het = TRUE, verbose = FALSE, model = "sarar")
summary(res)
effects <- impacts(res, n_mvn = 3000)
summary(effects)
}
