% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.impacts.sphet.R
\name{impacts.sphet}
\alias{impacts.sphet}
\title{describedIn impacts}
\usage{
\method{impacts}{sphet}(object, ...)
}
\arguments{
\item{object}{An object of class sphet}

\item{...}{Additional arguments to be passed}
}
\value{
Estimate of the Average Total, Average Direct, and Average Indirect Effects
}
\description{
describedIn impacts
}
\examples{
data(columbus, package="spdep")
listw <- spdep::nb2listw(col.gal.nb)
res <- spreg(CRIME~HOVAL + INC, data=columbus , listw= listw,
            het = TRUE, verbose = FALSE, model = "sarar")
summary(res)
effects <- impacts(res, n_mvn = 3000)
summary(effects)
}
