% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sphet.R
\name{print.sphet}
\alias{print.sphet}
\title{print method for class sphet}
\usage{
\method{print}{sphet}(x, digits = max(3, getOption("digits") - 3),...)
}
\arguments{
\item{x}{an object of class 'sphet'}

\item{digits}{minimal number of significant digits, see \code{print.default}}

\item{...}{additional arguments to be passed}
}
\description{
Method used to print objects of class \code{'summary.sphet'} and \code{'sphet'}
}
\details{
The summary function summary.sphet returns an objects of class 'sphet' 
organized in a coefficient matrix.
}
\examples{
library(spdep)
data(columbus)
listw <- nb2listw(col.gal.nb)
res <- spreg(CRIME~HOVAL + INC, data=columbus, listw=listw, model ="sarar")
summary(res)
}
\seealso{
\code{\link{gstslshet}}, \code{\link{stslshac}}
}
\author{
Gianfranco Piras\email{gpiras@mac.com}
}
