\name{sphet-package}
\alias{sphet-package}
\alias{sphet}

\title{Estimation of spatial models with heteroskedastic innovations}
\description{
A set of functions to estimate spatial models with heteroskedastic innovations
}

\references{

Piras, Gianfranco (2010) 
sphet: Spatial Models with Heteroskedastic Innovations in R,
\emph{Journal of Statistical Software} June 2010, Volume 35, Issue 1. 


Bivand, R; Millo, G; Piras, G. (2021)
A Review of Software for Spatial Econometrics in R
\emph{Mathematics} 9 (11):1276.
         
Bivand, R; Piras, G. (2015)
Comparing Implementations of Estimation Methods for Spatial Econometrics,
  \emph{Journal of Statistical Software}, Volume 63, Issue 18, 1--36.
  
}

\details{
\tabular{ll}{
Package: \tab sphet\cr
Type: \tab Package\cr
Version: \tab 1.12\cr
Date: \tab 2021-06-17\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{ Gianfranco Piras \email{gpiras@mac.com}}
