% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.impacts.sphet.R
\name{summary.impacts_sphet}
\alias{summary.impacts_sphet}
\title{Summary for spreg impacts}
\usage{
\method{summary}{impacts_sphet}(object, ...)
}
\arguments{
\item{object}{An object of class impacts.sphet}

\item{...}{Additional arguments that can be passed to the function}
}
\value{
Summary of the Average Total, Average Direct, and Average Indirect Effects
}
\description{
Summary for spreg impacts
}
\examples{
data(columbus, package = "spdep")
listw <- spdep::nb2listw(col.gal.nb)
res <- spreg(CRIME~HOVAL + INC, data=columbus , listw= listw,
            het = TRUE, verbose = FALSE, model = "sarar")
summary(res)
effects <- impacts(res, n_mvn = 3000)
summary(effects)
}
