% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impacts.R
\name{impacts}
\alias{impacts}
\title{Generate impacts for spreg lag and sarar models}
\usage{
impacts(obj, ...)
}
\arguments{
\item{obj}{An object of class sphet}

\item{...}{Additional arguments to be passed}
}
\value{
Estimate of the Average Total, Average Direct, and Average Indirect Effects
}
\description{
Generate impacts for spreg lag and sarar models
}
\examples{
data(columbus, package="spdep")
listw <- spdep::nb2listw(col.gal.nb)
res <- spreg(CRIME~HOVAL + INC, data=columbus , listw= listw,
            het = TRUE, verbose = FALSE, model = "sarar")
summary(res)
effects <- impacts(res, listw = listw, R = 399)
summary(effects)
}
