\name{sphet-package}
\alias{sphet-package}
\alias{sphet}

\title{Estimation of spatial models with heteroskedastic innovations}
\description{
A set of functions to estimate spatial models with heteroskedastic innovations
}

\references{

Piras, Gianfranco (2010) 
sphet: Spatial Models with Heteroskedastic Innovations in R,
\emph{Journal of Statistical Software} June 2010, Volume 35, Issue 1. 
}

\details{
\tabular{ll}{
Package: \tab sphet\cr
Type: \tab Package\cr
Version: \tab 1.0-0\cr
Date: \tab 2010-06-03\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{ Gianfranco Piras \email{gpiras@mac.com}}
