\name{sphet-package}
\alias{sphet-package}
\alias{sphet}

\title{Estimation of spatial models with heteroskedastic innovations}
\description{
A set of functions to estimate spatial models with heteroskedastic innovations
}
\details{
\tabular{ll}{
Package: \tab sphet\cr
Type: \tab Package\cr
Version: \tab 0.1.22\cr
Date: \tab 2009-12-17\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{ Gianfranco Piras \email{gpiras@mac.com}}
